+++ src/api/v2/order.rs
@@ -717,14 +717,6 @@ Endpoint! {
     /* 200 */ OK,
   ],
   Err => PostError, [
-    /// The order submission was not permitted. That can have multiple
-    /// reasons, including (but not necessarily limited to):
-    /// - not enough funds are available
-    /// - the order is of a certain order type that cannot be submitted
-    ///   at this time of day (e.g., market-open orders must be
-    ///   submitted after 7:00pm and before 9:28am and will be rejected
-    ///   at other times)
-    /* 403 */ FORBIDDEN => NotPermitted,
     /// Some data in the request was invalid.
     /* 422 */ UNPROCESSABLE_ENTITY => InvalidInput,
   ]
@@ -756,14 +748,6 @@ Endpoint! {
     /* 200 */ OK,
   ],
   Err => PatchError, [
-    /// The order change was not permitted. That can have multiple
-    /// reasons, including (but not necessarily limited to):
-    /// - not enough funds are available
-    /// - the order is of a certain order type that cannot be submitted
-    ///   at this time of day (e.g., market-open orders must be
-    ///   submitted after 7:00pm and before 9:28am and will be rejected
-    ///   at other times)
-    /* 403 */ FORBIDDEN => NotPermitted,
     /// No order was found with the given ID.
     /* 404 */ NOT_FOUND => NotFound,
     /// Some data in the request was invalid.
@@ -1067,8 +1051,7 @@ mod tests {
 
     // When an extended hours order is submitted between 6pm and 8pm,
     // the Alpaca API reports an error:
-    // > {"code":42210000,"message":"extended hours orders between 6:00pm
-    // >   and 8:00pm is not supported"}
+    // > {"message":"extended hours orders between 6:00pm and 8:00pm is not supported"}
     //
     // So we need to treat this case specially.
     let result = test(true).await;