+++ src/data/v2/last_quotes.rs
@@ -92,9 +92,6 @@ EndpointNoParse! {
     /// The provided symbol was invalid or not found or the data feed is
     /// not supported.
     /* 400 */ BAD_REQUEST => InvalidInput,
-    /// The request was not permitted. Possible reasons include usage of
-    /// the SIP feed without having the corresponding subscription.
-    /* 403 */ FORBIDDEN => NotPermitted,
   ]
 
   fn base_url() -> Option<Str> {