+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Unreleased
   - This variant is now used to signal a multitude of conditions,
     including certain order submission issues
 - Bumped minimum supported Rust version to `1.63`
+- Bumped `websocket-util` dependency to `0.12`
+- Bumped `tokio-tungstenite` dependency to `0.20`
 
 
 0.27.2